+++ OSXvnc-server/rdr/FdOutStream.h
@@ -37,11 +37,11 @@ namespace rdr {
     int getFd() { return fd; }
 
     void flush();
-    int length();
-    void writeBytes(const void* data, int length);
+    size_t length();
+    void writeBytes(const void* data, size_t length);
 
   private:
-    int overrun(int itemSize, int nItems);
+    size_t overrun(int itemSize, size_t nItems);
     int fd;
     int bufSize;
     int offset;
