+++ OSXvnc-server/rdr/MemOutStream.h
@@ -40,13 +40,13 @@ namespace rdr {
       delete [] start;
     }
 
-    void writeBytes(const void* data, int length) {
+    void writeBytes(const void* data, size_t length) {
       check(length);
       memcpy(ptr, data, length);
       ptr += length;
     }
 
-    int length() { return ptr - start; }
+    size_t length() { return ptr - start; }
     void clear() { ptr = start; };
     void reposition(int pos) { ptr = start + pos; }
 
@@ -59,7 +59,7 @@ namespace rdr {
     // overrun() either doubles the buffer or adds enough space for nItems of
     // size itemSize bytes.
 
-    int overrun(int itemSize, int nItems) {
+    size_t overrun(int itemSize, size_t nItems) {
       int len = ptr - start + itemSize * nItems;
       if (len < (end - start) * 2)
         len = (end - start) * 2;
